[Feature #111504652] Extract repeated codes and validators to separate files
