+++ controllers/EntriesController.js
@@ -1,43 +1,19 @@
-import promise from 'bluebird';
-import Joi from 'joi';
-import jwt from 'jsonwebtoken';
-
-const options = {
-  // Initialization Options
-  promiseLib: promise
-};
-const pgp = require('pg-promise')(options);
-const connectionString = 'postgres://zehxatan:EP22Gmb985sp3eXwp2z94Hz-9nlbr4D4@stampy.db.elephantsql.com:5432/zehxatan';
-const db = pgp(connectionString);
-
-const validateEntry = (entry) => {
-  const schema = {
-    title: Joi.string().min(3).trim().required(),
-    body: Joi.string().min(3).trim().required(),
-    userid: Joi.number().min(1)
-  };
-  return Joi.validate(entry, schema);
-};
+import db from '../middleware/connectdb';
+import validateEntry from '../middleware/validateEntry';
 
 
 class EntriesController {
   // get all entries
   static async getAllEntries(req, res) {
-    const token = req.header('x-auth-token');
-    if (!token) return res.status(401).json({ message: 'Access denied, no token provided', status: 'Failed' });
-    const decoded = jwt.verify(token, 'oiraid');
-    req.user = decoded;
-    const entries = await db.any('SELECT * FROM entries where userid = $1', decoded.id);
+    const entries = await db.any('SELECT * FROM entries where userid = $1', req.user.id);
+    if (entries.length === 0) return res.status(404).json({ message: 'No entries posted yet', status: 'error' });
     res.status(200).json({ status: 'success', entries, message: 'Retrieved ALL Entries' });
   }
 
   // Get single entry
   static async getEntry(req, res) {
-    const token = req.header('x-auth-token');
-    if (!token) return res.status(401).json({ message: 'Access denied, no token provided', status: 'Failed' });
-    const decoded = jwt.verify(token, 'oiraid');
     const entryID = parseInt(req.params.id, 10);
-    const entry = await db.any('SELECT * FROM entries where id = $1 and userid =$2', [entryID, decoded.id]);
+    const entry = await db.any('SELECT * FROM entries where id = $1 and userid =$2', [entryID, req.user.id]);
     if (entry.length === 0) return res.status(404).json({ message: 'Entry does not exist', status: 'error' });
     res.status(200).json({ status: 'success', entry, message: 'Retrieved ONE entry' });
   }
@@ -46,10 +22,7 @@ class EntriesController {
   static async addEntry(req, res) {
     const { error } = validateEntry(req.body);
     if (error) return res.status(400).json({ message: error.details[0].message, status: 'Failed' });
-    const token = req.header('x-auth-token');
-    if (!token) return res.status(401).json({ message: 'Access denied, no token provided', status: 'Failed' });
-    const decoded = jwt.verify(token, 'oiraid');
-    req.body.userid = decoded.id;
+    req.body.userid = req.user.id;
     await db.any('insert into entries(userid, title, body) values(${userid}, ${title}, ${body})', req.body);
     res.status(201).json({ status: 'success', message: 'Inserted one Entry' });
   }
@@ -58,10 +31,7 @@ class EntriesController {
   static async updateEntry(req, res) {
     const { error } = validateEntry(req.body);
     if (error) return res.status(400).json({ message: error.details[0].message, status: 'Failed' });
-    const token = req.header('x-auth-token');
-    if (!token) return res.status(401).json({ message: 'Access denied, no token provided', status: 'Failed' });
-    const decoded = jwt.verify(token, 'oiraid');
-    const result = await db.result('update entries set title=$1, body=$2 where id=$3 and userid=$4', [req.body.title, req.body.body, parseInt(req.params.id, 10), decoded.id]);
+    const result = await db.result('update entries set title=$1, body=$2 where id=$3 and userid=$4', [req.body.title, req.body.body, parseInt(req.params.id, 10), req.user.id]);
     if (result.rowCount === 0) return res.status(404).json({ message: 'Entry does not exist', status: 'error' });
     res.status(200).json({ status: 'success', message: 'Updated one entry' });
   }
@@ -69,10 +39,7 @@ class EntriesController {
   // remove entry
   static async removeEntry(req, res) {
     const entryID = parseInt(req.params.id, 10);
-    const token = req.header('x-auth-token');
-    if (!token) return res.status(401).json({ message: 'Access denied, no token provided', status: 'Failed' });
-    const decoded = jwt.verify(token, 'oiraid');
-    const result = await db.result('delete from entries where id = $1 and userid= $2', [entryID, decoded.id]);
+    const result = await db.result('delete from entries where id = $1 and userid= $2', [entryID, req.user.id]);
     if (result.rowCount === 0) return res.status(404).json({ message: 'Entry does not exist', status: 'error' });
     res.status(200).json({ status: 'success', message: 'Entry deleted successfully' });
   }
