+++ middleware/validateSignIn.js
@@ -1,0 +1,12 @@
+import Joi from 'joi';
+const validateSignIn = (user) => {
+  const schema = {
+    email: Joi.string().min(5).max(255).required()
+      .email(),
+    password: Joi.string().min(5).max(255).required()
+
+  };
+  return Joi.validate(user, schema);
+};
+
+export default validateSignIn;
